# PERF: put some Timetamp methods in _Timestamp