```
In [2]: ts = pd.Timestamp.now()                                                                                                                                                                                    

In [3]: %timeit ts.is_month_start                                                                                                                                                                                  
58.3 ns ± 1.5 ns per loop (mean ± std. dev. of 7 runs, 10000000 loops each)  # <-- PR
82.5 ns ± 2.16 ns per loop (mean ± std. dev. of 7 runs, 10000000 loops each)  # <-- master
```